deviate from Hyku's default metadata as little as possible
We were following PALs knapsack's example of solving the issues with the
based_near field. However, their approach of overriding the
basic_metadata to only include the based_near field was based on their
need to customize many fields across most of the works. Since HykuUP
tries to stick as close to base Hyku as possible, this commit opts for a
simpler approach that necessitates less overrides