⬆️ Update submodule to 6448afb
In v6.2 when flexible=false, the controlled vocabulary fields rendered a
text field in the new/edit forms rather than a dropdown. This allowed
the user to enter data that did not match a predefined mapping
(e.g. `term: "In Copyright"`) defined in the hyku/config/authorities yml
files. Controlled vocabulary maps the term from the form dropdown to the
id that is persisted
(e.g. `id: http://rightsstatements.org/vocab/InC/1.0/`). The incorrect
data caused a KeyError the catalog search resulting in the application
breaking.

This commit adds hardcoded mappings from the form property to the
controlled vocabulary service key. Most of keys are the pluralized
version of the form property but since that is not always the case, it
was decided that hardcoding would prevent possible errors.

Ref:
- https://github.com/notch8/hykuup_knapsack/issues/531